requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
Jinja2==2.8
MarkupSafe==0.23
SQLAlchemy==1.0.11
WTForms==2.1
Werkzeug==0.11.3
argparse==1.2.1
cycler==0.9.0
itsdangerous==0.24
matplotlib==1.5.1
numpy==1.10.4
pandas==0.17.1
pyparsing==2.0.7
pyserial==3.0.1
python-dateutil==2.4.2
pytz==2015.7
six==1.10.0
wsgiref==0.1.2
